+++ example/tests/api/endpoints/endpoints.ts
@@ -1,0 +1,5 @@
+export const postEndpoints = {
+    getPosts: '/posts',
+    getPost: '/posts/1', // Example endpoint to fetch a specific post
+    invalidEndpoint: '/invalidEndpoint', // Example of an invalid endpoint
+};
