--- example/tests/api/rest/example.config.ts
@@ -1,6 +1,0 @@
-import { ApiTestsConfig } from '../../../../src/api/config/types';
-
-export const apiTestsConfig: ApiTestsConfig = {
-    restApiUrl: "https://jsonplaceholder.typicode.com",
-    // Add other configurations if needed
-};
