+++ example/tests/api/rest/posts.spec/posts.spec.ts
@@ -1,8 +1,9 @@
 import { expect } from 'chai';
 import { RestApiClientHelper } from '../../../../../src/api/rest/RestApiClientHelper';
 import { apiTestsConfig } from '../../config/example.config';
-import { Post } from '../../types/apiTypes';
-import { postEndpoints, postResponseSchema } from './testData/posts.testdata';
+import { Post } from '../../types/requestTypes';
+import { postResponseSchema, postsResponseSchema } from '../../types/responseTypes';
+import { postEndpoints } from '../../endpoints/endpoints'
 
 describe("REST API Posts Tests", () => {
     const restApiUrl = apiTestsConfig.restApiUrl;
@@ -12,13 +13,13 @@ describe("REST API Posts Tests", () => {
     const client = new RestApiClientHelper(restApiUrl);
 
     it("should fetch posts correctly", async () => {
-        const posts: Post[] = await client.sendRequest('get', postEndpoints.getPosts, undefined, postResponseSchema);
+        const posts: Post[] = await client.sendRequest('get', '/posts', undefined, postsResponseSchema);
         expect(posts).to.be.an('array');
         expect(posts.length).to.be.greaterThan(0);
     });
 
     it("should fetch a specific post", async () => {
-        const post = await client.sendRequest('get', postEndpoints.getPost);
+        const post = await client.sendRequest('get', '/posts/1', undefined, postResponseSchema);
         expect(post).to.be.an('object');
         expect(post).to.have.property('id', 1);
     });
