--- example/tests/api/types/graphqlTypes.ts
@@ -1,12 +1,0 @@
-export type Country = {
-  code: string;
-  name: string;
-};
-
-export type CountryData = {
-  country: Country;
-};
-
-export type CountriesData = {
-  countries: Country[];
-};
